requirements.txt
Flask>=2.1.0
mysql-connector-python>=8.0.0
openai>=1.0.0
python-dotenv>=1.0.0